Refactor bindingutilities and outputhelper into .cpp and header files
